型修正

+++ scr/CSVLoad/CardLoader.h
@@ -45,7 +45,7 @@ public:
             data.cost = std::stoi(cols[2]);
             int actionTypeValue = std::stoi(cols[3]);
             data.actionType = static_cast<ActionType>(actionTypeValue);
-            data.power = std::stoi(cols[4]);
+            data.power = std::stof(cols[4]);
             int targetTypeValue = std::stoi(cols[5]);
             data.targetType = static_cast<TargetType>(targetTypeValue);
             data.turn = std::stoi(cols[6]);
